+++ TimberCommons/CommonUIPatches/GoodConsumingBuildingFragmentPatch.cs
@@ -4,6 +4,7 @@
 
 using System;
 using HarmonyLib;
+using IgorZ.TimberCommons.Common;
 using IgorZ.TimberDev.UI;
 using Timberborn.GoodConsumingBuildingSystem;
 using Timberborn.GoodConsumingBuildingSystemUI;
@@ -20,8 +21,22 @@ namespace IgorZ.TimberCommons.CommonUIPatches {
 /// </remarks>
 [HarmonyPatch(typeof(GoodConsumingBuildingFragment), nameof(GoodConsumingBuildingFragment.UpdateProgressBar))]
 static class GoodConsumingBuildingFragmentPatch {
-  const string SupplyRemainingLocKey = "GoodConsuming.SupplyRemaining";
-  static string _localizedSupplyRemainingTmpl;
+  const string NoTilesToIrrigateLocKey = "IgorZ.TimberCommons.WaterTower.NoTilesToIrrigate";
+
+  // ReSharper disable once UnusedMember.Local
+  static bool Prefix(bool __runOriginal, ILoc ____loc,
+                     Timberborn.CoreUI.ProgressBar ____hoursLeftBar, Label ____hoursLeft,
+                     GoodConsumingBuilding ____goodConsumingBuilding) {
+    if (!__runOriginal) {
+      return false;  // The other patches must follow the same style to properly support the skip logic!
+    }
+    if (____goodConsumingBuilding.GoodPerHour > float.Epsilon) {
+      return true;
+    }
+    ____hoursLeftBar.SetProgress(0);
+    ____hoursLeft.text = ____loc.T(NoTilesToIrrigateLocKey);
+    return false;
+  }
 
   // ReSharper disable once UnusedMember.Local
   static void Postfix(bool __runOriginal, ILoc ____loc, Label ____hoursLeft,
@@ -29,13 +44,10 @@ static class GoodConsumingBuildingFragmentPatch {
     if (!__runOriginal) {
       return;  // The other patches must follow the same style to properly support the skip logic!
     }
-    if (_localizedSupplyRemainingTmpl == null) {
-      var original = ____loc.T(SupplyRemainingLocKey, "###");
-      _localizedSupplyRemainingTmpl = original.Substring(0, original.IndexOf("###", StringComparison.Ordinal)) + "{0}";
+    if (!Features.GoodConsumingBuildingUIDaysHoursForAll) {
+      return;
     }
-    ____hoursLeft.text = string.Format(
-        _localizedSupplyRemainingTmpl,
-        CommonFormats.DaysHoursFormat(____loc, ____goodConsumingBuilding.HoursUntilNoSupply));
+    ____hoursLeft.text = CommonFormats.FormatSupplyLeft(____loc, ____goodConsumingBuilding.HoursUntilNoSupply);
   }
 }