+++ TimberCommons/package/CHANGELOG.md
@@ -3,6 +3,7 @@
 * [Fix #43] Water towers override the highlighted range of other buildings.
 * [Fix #45] Mechanical towers don't properly update the effective range.
 * [Enhancement] Improved powered irrigation towers performance when the Power Grid gets fluctuations.
+* [Enhancement] Update `deDE` localization (by @juf0816).
 
 # v1.6 (21 Feb 2024):
 * [Fix #33] Water valve doesn't handle contamination correctly.