+++ TimberCommons/package/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v1.7 (started on 3/1/2024):
+* [Fix #37] Water tower with zero coverage consumes water.
+* [Fix #43] Water towers override the highlighted range of other buildings.
+* [Fix #45] Mechanical towers don't properly update the effective range.
+* [Fix #48] Irrigation range highlighting from the previous saved game says after loading.
+* [Fix #49] Manufacture water tower misses localization string.
+* [Enhancement] Improved powered irrigation towers performance when the Power Grid gets fluctuations.
+* [Enhancement] Update `deDE` localization (by @juf0816).
+* [Enhancement] Prevent beavers contamination when moving in tunnels under bad water.
+
 # v1.6 (21 Feb 2024):
 * [Fix #33] Water valve doesn't handle contamination correctly.
 * [Fix #38] "Supply will last for" text disappeared in UI.